+++ appliances/flower_supernode/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the Flower SuperNode appliance will be documented in this file.
 
-## [1.25.0-1.0.0] - 2026-06-18
+## [1.31.0-1.0.0] - 2026-06-18
 
 ### Added